requirements.txt
Pillow~=7.2.0
tensorflow-gpu~=2.3.0
numpy~=1.19.1